pyproject.toml
[project]
name = "cprompt"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Paul Bailey", email = "paul@neutron.studio"},
]
dependencies = [
    "django>=5.0.6",
    "django-ninja>=1.1.0",
    "pytest-django>=4.8.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
_.env_file = ".env"
dev = "python manage.py runserver 0.0.0.0:8000"
test = "pytest -sx"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "cprompt.project.settings"
python_files = ["tests.py", "test_*.py", "*_test.py", "testing/python/*.py"]
